backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "torch>=2.9.1",
    "transformers",
    "uvicorn>=0.38.0",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.4.0",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }
